+++ src/System.CommandLine.CommandHandler/FactoryParameter.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace System.CommandLine.CommandHandler
+{
+    public class FactoryParameter : Parameter
+    {
+        public ITypeSymbol FactoryType { get; }
+        public string LocalName { get; }
+        private string ParameterName => LocalName.ToLowerInvariant();
+
+        public FactoryParameter(RawParameter rawParameter, ITypeSymbol factoryType)
+            : base(rawParameter.ValueType)
+        {
+            LocalName = rawParameter.LocalName;
+            FactoryType = factoryType;
+        }
+
+        public override string GetValueFromContext() => "";
+
+        public override string GetPropertyDeclaration()
+            => $"private {FactoryType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)} {LocalName} {{ get; }}";
+
+        public override string GetPropertyAssignment()
+            => $"{LocalName} = {ParameterName};";
+
+        public override (string Type, string Name) GetMethodParameter()
+            => (FactoryType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat), ParameterName);
+    }
+}
